+++ tailwind.config.cjs
@@ -2,7 +2,7 @@
 module.exports = {
 	darkMode: 'media',
 	content: [
-		'./src/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue}',
+		'./src/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue,json}',
 		'./node_modules/flowbite/**/*.js'
 	],
 	theme: {